fix compilation ; removed some duplicate code
