+++ bevy_rapier3d/Cargo.toml
@@ -61,6 +61,7 @@ bevy = { version = "0.14", default-features = false, features = [
 approx = "0.5.1"
 glam = { version = "0.27", features = ["approx"] }
 divan = "0.1"
+bevy_rapier_benches3d = { version = "0.1", path = "../bevy_rapier_benches3d" }
 
 [package.metadata.docs.rs]
 # Enable all the features when building the docs on docs.rs
